Drop v332 migration, keep project_board.sorting as int8
Per review feedback, the 127-column cap is intentional (maxProjectColumns
is 20), so the DB schema is left as-is and no migration is needed. Reverts
the Column.Sorting widening to match.

+++ routers/api/v1/repo/project.go
@@ -574,7 +574,7 @@ func EditProjectColumn(ctx *context.APIContext) {
 		column.Color = *form.Color
 	}
 	if form.Sorting != nil {
-		column.Sorting = *form.Sorting
+		column.Sorting = int8(*form.Sorting)
 	}
 
 	if err := project_model.UpdateColumn(ctx, column); err != nil {
